==> 1- Get Data from Inside Schools.ipynb <==
## Inside Schools Data Collection

import pandas as pd
import requests
from bs4 import BeautifulSoup
import dateutil.parser
import re
from time import sleep



Read List of school IDs from downloaded directory

school_nums = pd.read_csv('/Users/adamsantos/ds/metis/Regression School Project/2018_DOE_High_School_Directory.csv')
school_nums = school_nums['dbn']
school_codes = list(school_nums.values)
tester = school_nums[1]
#tester
#school_codes

Test connection

url = 'https://insideschools.org/school/14K454'

response = requests.get(url)

response.status_code

page = response.text
soup = BeautifulSoup(page, "lxml")

Function to clean strings

def clean_me(string):
    string = string.strip('\n')
    string = string.rstrip('\n')
    if string and string[-1] == '%':
        string = string.replace('%', '')
    try:
        string = float(string)
    except:
        string = string
    return string

Function to get data for a single school from soup object

def catch_em_all(soup, school_code):
    
    #Get base statistics
    titles = soup.find_all('div',{'class' : 'stat-title'})
    titles_final = [clean_me(x.text) for x in titles]
    
    values = soup.find_all('div',{'class' : 'stat-value'})
    values_final = [clean_me(x.text) for x in values]
    
    values_final = [x for x in values_final if not str(x).startswith('This school offers')]
    
    #Get teacher rating
    stars = soup.find_all('span',{'class' : 'stat-star icon-star gold'})
    rating = len(stars)
    titles_final.append('Teacher Effectiveness')
    values_final.append(rating)
    
    #Get school address
    address = soup.find('div',{'class' : 'location-address'}).text
    address = address.strip('\n')
    address = address.replace('\n',' ')
    address = ' '.join(address.split())
    titles_final.append('Address')
    values_final.append(address)
    
    #Put school code into final dict
    titles_final.append('School Code')
    values_final.append(school_code)
    
    name = soup.find('h1',{'class' : 'school-name'}).text
    name=name.strip('\n')
    name=name.rstrip('\n')
    titles_final.append('School Name')
    values_final.append(name)
    
    dictionary = dict(zip(titles_final, values_final))
    return dictionary


#test = catch_em_all(soup,'02M560')
#test

Get data from each school 

data=[]

for site in school_codes:
    url = 'https://insideschools.org/school/'+site
    response = requests.get(url)
    
    if response.status_code == 200:
        page = response.text
        soup = BeautifulSoup(page, "lxml")
    
        data.append(catch_em_all(soup,site))
    
    sleep(0.5)
  

df=pd.DataFrame(data)

df=df.dropna(thresh=100, axis=1)

df.head()

Save for future use

df.to_pickle("oct_3.pkl")
