==> 3- Combine Inside Schools and Funding.ipynb <==
## 3- Combine data from Inside Schools and Chalkbeat

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
%matplotlib inline
import seaborn as sns
from scipy.stats.stats import pearsonr 

Load data from each website

df_is = pd.read_pickle("oct_3.pkl")
df_fund = pd.read_pickle("funding_per_student.pkl")

Set column names of Inside Schools dataframe

df_is = df_is[['How many students graduate in 4 years?','School Name','School Code','Address', 'Asian', 'Average daily attendance', 'Black',
       'English language learners', 'Enrollment',
       'Free or reduced priced lunch', 'Hispanic',
       'How many English language learners graduate in 4 years?',
       'How many graduates earn Advanced Regents diplomas?',
       'How many graduates stay enrolled in college for at least 3 semesters?',
       'How many students graduate with test scores high enough to enroll at CUNY without remedial help?',
       'How many students miss 18 or more days of school?',
       'How many students say they feel safe in the hallways, bathrooms and locker rooms?',
       'How many students take a college-level course or earn a professional certificate?',
       'How many students think bullying happens most or all of the time at this school?',
       'How many students were suspended?',
       'How many students with disabilities graduate in 4 years?',
       'How many teachers have 3 or more years of experience teaching?',
       'How many teachers say order and discipline are maintained at this school?',
       'How many teachers say the principal is an effective manager?',
       'How many teachers say they would recommend this school to other families?',
       'Metal detectors?', 'Other',  
       'Shared campus?', 'Students with disabilities', 'Teacher Effectiveness',
       'White', 'Years of principal experience at this school']]


Use close matches to change school names that differ slightly to match them between dataframes

import difflib

school_names=df_is['School Name'].map(lambda x: difflib.get_close_matches(x, df_fund['School'], n=1))
school_names = [s[0] if s else '' for s in school_names]
len(school_names)

Merge data frames to put inside school data and funding data together

df_merged = pd.merge(df_is,df_fund,left_on='New School Name',right_on='School',how='left')
#df_merged.head()

### Clean data for use in Linear Regression

Drop schools that don't have the target value available

df_merged = df_merged[pd.notnull(df_merged['How many students graduate in 4 years?'])]
#df_merged.info()

Drop data related to metal detectors, shared campuses, and students with disabilities due to lack of data (less than half of schools have this information)

df_merged=df_merged.drop(['Metal detectors?'], axis=1)

df_merged=df_merged.drop(['Shared campus?'], axis=1)

df_merged=df_merged.drop(['How many students with disabilities graduate in 4 years?'], axis=1)

Drop redundant School Name column

df_merged=df_merged.drop(['School'], axis=1)

Fill in missing values for a few schools that were missing information

df_merged['How many graduates stay enrolled in college for at least 3 semesters?'] = df_merged['How many graduates stay enrolled in college for at least 3 semesters?'].fillna(df_merged['How many students graduate with test scores high enough to enroll at CUNY without remedial help?'])

df_merged['How many students were suspended?'].fillna((df_merged['How many students were suspended?'].mean()), inplace=True)


Convert school funding column to float

df_merged['$ per Student'] = df_merged['$ per Student'].str.replace(',','')
df_merged['$ per Student'] = df_merged['$ per Student'].astype(float)
df_merged.head()

Save clean dataframe for use in Linear Regression

df_merged.to_pickle("clean_school_df.pkl")
